完成

+++ app/Http/Controllers/DiaryController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 
 use App\Diary;
 use App\Http\Requests\CreateDiary;
+use Illuminate\Support\Facades\Auth;
 
 class DiaryController extends Controller
 {
@@ -18,7 +19,9 @@ class DiaryController extends Controller
         //diariesデーブルのデータを全件取得する。
         //取得した結果を画面で確認。
 
-        $diaries = Diary::all();
+        // $diaries = Diary::all();
+
+        $diaries = Diary::with('likes')->orderBy('id','desc')->get();
         // dd($diaries);
         //dd();var_dump と dieが同時に実行される
         //views/diaries/index.blade.phpを表示
@@ -40,9 +43,11 @@ class DiaryController extends Controller
         $diary = new Diary();
 
         //Diaryモデルを使って、DBに日記を保存
-       //$diary->カラム名　= カラム　に設定したい値
+       //$diary->カラム名 = カラム に設定したい値
         $diary->title = $request->title;
         $diary->body = $request->body;
+
+        $diary->user_id = Auth::user()->id;
 
         //DBに保存実行
         $diary->save();
@@ -54,12 +59,94 @@ class DiaryController extends Controller
     //日記を削除する為のメソッド
     public function destroy (int $id)
     {
-        
+
         //diaryモデルを使用してIDが一致する日記の取得
         $diary = Diary::find($id);
+        //ログインユーザーが日記の投稿者かチェックする
+    if(Auth::user()->id != $diary->user_id) {
+
+        abort(403);
+    }
         //取得した日記の削除
         $diary->delete();
         //一覧画面にリダイレクト
         return redirect()->route('diary.index');
+    }
+
+    //編集画面を表示する
+    public function edit(diary $diary)
+    {
+        //ログインユーザーが日記の投稿者かチェックする
+        if(Auth::user()->id != $diary->user_id) {
+
+            abort(403);
+        }
+        // 受け取ったIDを元に日記を取得する
+        // $diary = Diary::find($id);
+        //編集画面を返す。同時に画面に取得した日記を渡す。
+
+        return view('diaries.edit',[
+            'diary' => $diary
+            //.キー =>値
+        ]);
+    }
+    //日記を更新し、一覧画面にリダイレクトする
+    //-$id:編集対象の日記のID
+    //-$request:リクエストの内容。ここに画面で入力された文字が格納されている。
+
+    public function update(int $id, Creatediary $request)
+    {
+    //受け取ったIDを元に日記を取得
+    $diary = Diary::find($id);
+
+    //ログインユーザーが日記の投稿者かチェックする
+    if(Auth::user()->id != $diary->user_id) {
+
+        abort(403);
+    }
+    //取得した日記のタイトル、本文を書き換える
+    //Diaryモデルを使って、DBに日記を保存
+       //$diary->カラム名 = カラム　に設定したい値
+       $diary->title = $request->title;
+       $diary->body = $request->body;
+
+       //DBに保存実行
+       $diary->save();
+    
+    //一眼ページにリダイレクト
+    return redirect()->route('diary.index');
+
+
+
+    
+    }
+
+    //いいねが押された時の処理（まだ押されてない状態で）
+    public function like (int $id)
+    {
+        //いいねされた日記の取得
+        $diary = Diary::find($id);
+        //likes()は関数likes
+        //attach:多対多のデータを登録するメソッド
+        $diary->likes()->attach(Auth::user()->id);
+        //通信が成功したことを返す
+        return response()
+                        ->json(['success'=>'いいね完了']);
+
+
+
+    }
+    //いいねが押された時の処理（すでに押されている状態で）
+    public function dislike (int $id)
+    {
+        //いいね解除された日記の取得
+        $diary = Diary::find($id);
+        //likes()は関数likes
+        //attach:多対多のデータを削除するメソッド
+        $diary->likes()->detach(Auth::user()->id);
+
+        //通信が成功したことを返す
+        return response()
+                        ->json(['success'=>'いいね完了']);
     }
 }
